+++ src/azure-cli/HISTORY.rst
@@ -19,6 +19,7 @@ Release History
 
 * Fix issue #11658: `az group export` command does not support `--query` and `--output` parameters
 * Fix issue #10279: The exit code of `az group deployment validate` is 0 when the verification fails
+* Add new parameter `--managed-by` to support adding managedBy information for command `az group create`.
 
 **IoT Central**